pyproject.toml
[project]
name = "myproject"
version = "0.1.0"
description = "My project description"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django>=5.1.2,<5.2",
    "django-stubs-ext>=5.1.1,<5.2",
    "django-debug-toolbar>=4.4.6",
    "django-extensions>=3.2.3",
    "djangorestframework>=3.15.2",
    "markdown>=3.7",
    "django-filter>=24.3",
    "drf-spectacular>=0.27.2",
    "django-health-check>=3.18.3",
    "django-environ>=0.11.2",
]

[project.optional-dependencies]
production = [
    "uvicorn>=0.32.0",
    "uvloop>=0.21.0; platform_system != 'Windows'",
]

[dependency-groups]
dev = [
    "django-stubs[compatible-mypy]>=5.1.1,<5.2",
    "ipython>=8.29.0",
    "mypy>=1.13.0,<1.14",
    "ptpython>=3.0.29",
    "ruff>=0.7.2",
    "werkzeug>=3.1.3",
]

[project.scripts]
appsite-manage = "myproject:manage"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["myproject"]

[tool.hatch.build.targets.sdist]
include = [
    "/myproject/**/*.py",
    "/myproject/**/templates/**/*",
    "/myproject/**/static/**/*",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.format]
docstring-code-format = true

[tool.ruff.lint]
# Enable flake8-bugbear (`B`) rules, in addition to the defaults.
select = ["E4", "E7", "E9", "F", "B", "I", "N", "UP", "DJ", "LOG", "SLF"]
# Avoid enforcing line-length violations (`E501`)
ignore = ["E501"]
# Avoid trying to fix flake8-bugbear (`B`) violations.
unfixable = ["B"]

[tool.ruff.lint.per-file-ignores]
# Ignore `E402` (import violations) in all `__init__.py` files, and in selected subdirectories.
"__init__.py" = ["E402"]
# Ignore star imports in Django settings modules
"myproject/settings.py" = ["F403", "F405"]

[tool.mypy]
plugins = ["mypy_django_plugin.main"]

[tool.django-stubs]
django_settings_module = "myproject.settings"
